Moved mpi testing cleanup into conftest fixture
